requirements.txt
# Dependencias comunes
pika
psycopg2-binary

# Dependencias para la API
fastapi
uvicorn[standard]
